tools: Fix source code path in analyze.sh
The analyze.sh (in case the bug report) of the Wind (operating system) and Tempest (the kernel) was
simply not working because the source code path (in case it was "./src") was not the correct path for
source code and the error:

./file.c:??-??: fatal error: 'file.h' file not found
   13 | #include "file.h"
      |          ^~~~~~~~~~~~~~~~~~~~~

This is not the behavior of the original code. So the fix is a simple fix. Change "./src" to "./kernel"

Fixes: f817963cf9a948dc78268eca026eeea758265d66 ("Finished organization on source code tree")
Signed-off-by: Russian95CrE <russian95@tempestfoundation.org>