shell: Fix `test_graphics()` by clearing the screen
Before, `cmd_test_graphics()` didn't called a clear function.
That's ok, but when I added `cmd_test_graphics()` to call
a clear function. That fucks that result on the end of
`ksleep()`. This commit fixes this by adding a simple and
easy task of... adding `kvideo_clear()` to the end of
`cmd_test_graphics()`.

Fixes: c5530298673e06f85e8c65378e94be2b0ba4d793 ("Use kis_video_ready() in test_graphics")
Signed-off-by: Russian95CrE <russian95@tempestfoundation.org>