docs: Remove NAVO as it is a discarded idea
Today (30/08/2025), I will discard the idea of the
NAVO project.  It is stupid and not useful for now in
present.  DOS is useless, it is history, it was good
back then, but horrible for daily use now. Please...
Just forever don't remember this stupid idea I got.
This commit fixes that.

Signed-off-by: Russian95CrE <russian95@tempestfoundation.org>