docs: Fix query path in `README.md`
Some days ago, I changed the repo name from `wind.git` to
`wt-src.git` because this repo includes Wind (the OS) and
also the Tempest (the kernel).  This project is mono-repo.
But... I did forgot to fix the README.md which used the
`wind.git` to request the star history and contrib.rocks.
This commit fixes that.

Signed-off-by: Russian95CrE <russian95@tempestfoundation.org>